Provide more control over factory creation and failure handling
Update `SpringFactoriesLoader` so that factory classes constructor
parameters cam be resolved by type and failures can be handled on a
per-instance basis.

This update allows the creation of any factory class that has a
single non-private constructor. Constructor parameters are resolved
by type using the `ParameterResolver` interface. A number of
convenience methods are provided to allow resolvers to build
built. For example:

	ParameterResolver.of(String.class, "tests")
		.and(Integer.class, 123);

In addition, a new `FactoryInstantiationFailureHandler` has been
provided that allows instantiation failures to be handled on a
per-factory bases. For example, to log trace messages for only
factories that can't be created the following can be used:

	FactoryInstantiationFailureHandler.logging(logger);

Closes gh-

Co-authored-by: Madhura Bhave <bhavem@vmware.com>
Co-authored-by: Andy Wilkinson <wilkinsona@vmware.com>